requirements.txt
# Web & Extraction
requests
beautifulsoup4
trafilatura

# AI & NLP
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
mistralai
pydantic

# UI & Data
streamlit
pandas
python-dotenv

# Export & Reporting
fpdf